genmakefile.sh: Change echo to printf for better portability & control
Solaris has too many echoes, some which support "-n" to not print newlines,
others which simply treat it as another item to echo to stdout.

The printf command is more portable, allows us better control over newline
insertion and quoting, and easier to read indicators for tabs.

After this change, when run on Solaris, the Makefile.cursors produced in
the redglass & whiteglass directories by genmakefile.sh are identical to
the versions currently checked into git.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>