genmakefile.sh: fix autogeneration of handhelds/Makefile.cursor
Previously handhelds/Makefile.cursor was manually fixed up after running
genmakefile.sh, in order to get the cursor image files out of the redglass
directory. This fixes the script to correctly generate the Makefile.cursor
files without human intervention.

After this change, the Makefile.cursor files produced in the redglass &
whiteglass directories by genmakefile.sh are still identical to the
versions currently checked into git.

The Makefile.cursors in the handhelds directory is mostly identical to
what was in git - it has lost the comments about needing human edits,
and it has lost the dependencies on the pngs for the larger size cursors
that aren't used in the handhelds cursor theme .cfg files (which just use
the *-16.png sizes of redglass).   The newly generated version is included
in this commit.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>